Remove configuration, bit of testing, logging
This commit removes configuration from this repository so that it can be
published publicly. Configuration of subnets and security groups is now
done through environment variables. I didn't find a way to populate
lists, so it's hardcoded to one subnet/SG at the moment, update
serverless.yml to change this.

Added some light testing through pep8 and a lambda that checks
connectivity to the Centreon instance, which can be invoked locally with
`sls invoke -f test_connectivity`.

The logging has been redone to output messages in JSON and should be
mostly machine parsable now.