Add logic to remove .env before deployment
At the moment, to prevent settings from previous deploys carrying over
you need to clean the whole environment, which slows down deployments.
This change will trigger a `rm -f .env` before deployment, forcing the
regeneration of the .env file and saving time if the deploy stage
happens to be on the same instance as the build stage.